+++ src/conversion/mesh_data/from_disk.rs
@@ -54,12 +54,12 @@ fn create_disk_plate(
             let i2 = ix1 + (iv + 1);
             let i3 = ix0 + (iv + 1);
 
-            indices.push(i0 as i32);
-            indices.push(i1 as i32);
-            indices.push(i2 as i32);
-            indices.push(i0 as i32);
-            indices.push(i2 as i32);
-            indices.push(i3 as i32);
+            indices.push(i0);
+            indices.push(i1);
+            indices.push(i2);
+            indices.push(i0);
+            indices.push(i2);
+            indices.push(i3);
         }
     }
 
@@ -115,12 +115,12 @@ fn create_disk_cylinder(radius: f32, height: f32, flip: bool, vv: f32, udiv: i32
             let i2 = ix1 + (iv + 1);
             let i3 = ix0 + (iv + 1);
 
-            indices.push(i0 as i32);
-            indices.push(i1 as i32);
-            indices.push(i2 as i32);
-            indices.push(i0 as i32);
-            indices.push(i2 as i32);
-            indices.push(i3 as i32);
+            indices.push(i0);
+            indices.push(i1);
+            indices.push(i2);
+            indices.push(i0);
+            indices.push(i2);
+            indices.push(i3);
         }
     }
